requirements.txt
# OMRChecker - Core Dependencies
# ================================

# Deep merge for configuration handling
deepmerge>=1.1.0

# Dot notation dictionary access
dotmap>=1.3.30

# JSON Schema validation
jsonschema>=4.17.3

# Plotting and visualization
matplotlib>=3.7.1

# Numerical computing
numpy>=1.25.0

# Data handling and analysis
pandas>=2.0.2

# Excel file support
openpyxl>=3.1.0

# Rich console output
rich>=13.4.2

# Screen info for display handling
screeninfo>=0.8.1

# Image processing
opencv-python>=4.8.0
pillow>=10.0.0

# ================================
# Web Interface Dependencies
# ================================

# Flask web framework
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# Async support for Socket.IO
eventlet>=0.33.0

# Secure filename handling
werkzeug>=2.3.0

# Windows TWAIN scanner support
pytwain>=2.1.0; platform_system=="Windows"

# Windows API access (for printer name resolution)
pywin32>=306; platform_system=="Windows"

# Linux SANE scanner support (optional, uncomment if needed)
# python-sane>=2.9.1; platform_system=="Linux"

# Network scanner discovery via mDNS/Zeroconf (eSCL/AirScan protocol)
zeroconf>=0.131.0
